Infrastructure/automation-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
redis>=5.0.0
hiredis>=2.2.0
asyncpg>=0.28.0
smbus2>=2.4.0

